first restructure and modified gitignore
