--- 5_Report/quarto/index.aux
@@ -1,6 +1,0 @@
-\relax 
-\providecommand*\new@tpo@label[2]{}
-\providecommand \babel@aux [2]{\global \let \babel@toc \@gobbletwo }
-\@nameuse{bbl@beforestart}
-\providecommand\hyper@newdestlabel[2]{}
-\providecommand*\HyPL@Entry[1]{}
